pie_chart_transactions
refactor nama file dan mencoba konsumsi data transaksi

Update chart_page.dart
